+++ sdk/appconfiguration/app-configuration/recordings/node/appconfigurationclient_setconfigurationsetting/recording_accepts_operation_options.js
@@ -1,5 +1,0 @@
-let nock = require('nock');
-
-module.exports.hash = "79146c6240cfa8030bc2371c5048d2f7";
-
-module.exports.testInfo = {"uniqueName":{"setConfigTestNA":"setConfigTestNA158696685579202551"},"newDate":{}}