+++ sdk/avs/arm-avs/samples-dev/workloadNetworksDeletePublicIPSample.ts
@@ -15,10 +15,10 @@ import { DefaultAzureCredential } from "@azure/identity";
  * This sample demonstrates how to Delete a Public IP Block by id in a private cloud workload network.
  *
  * @summary Delete a Public IP Block by id in a private cloud workload network.
- * x-ms-original-file: specification/vmware/resource-manager/Microsoft.AVS/stable/2021-12-01/examples/WorkloadNetworks_DeletePublicIPs.json
+ * x-ms-original-file: specification/vmware/resource-manager/Microsoft.AVS/stable/2022-05-01/examples/WorkloadNetworks_DeletePublicIPs.json
  */
 async function workloadNetworksDeletePublicIP() {
-  const subscriptionId = "{subscription-id}";
+  const subscriptionId = "00000000-0000-0000-0000-000000000000";
   const resourceGroupName = "group1";
   const publicIPId = "publicIP1";
   const privateCloudName = "cloud1";