+++ sdk/avs/arm-avs/samples/v3/javascript/placementPoliciesUpdateSample.js
@@ -15,15 +15,17 @@ const { DefaultAzureCredential } = require("@azure/identity");
  * This sample demonstrates how to Update a placement policy in a private cloud cluster
  *
  * @summary Update a placement policy in a private cloud cluster
- * x-ms-original-file: specification/vmware/resource-manager/Microsoft.AVS/stable/2021-12-01/examples/PlacementPolicies_Update.json
+ * x-ms-original-file: specification/vmware/resource-manager/Microsoft.AVS/stable/2022-05-01/examples/PlacementPolicies_Update.json
  */
 async function placementPoliciesUpdate() {
-  const subscriptionId = "{subscription-id}";
+  const subscriptionId = "00000000-0000-0000-0000-000000000000";
   const resourceGroupName = "group1";
   const privateCloudName = "cloud1";
   const clusterName = "cluster1";
   const placementPolicyName = "policy1";
   const placementPolicyUpdate = {
+    affinityStrength: "Must",
+    azureHybridBenefitType: "SqlHost",
     hostMembers: [
       "fakehost22.nyc1.kubernetes.center",
       "fakehost23.nyc1.kubernetes.center",
@@ -31,8 +33,8 @@ async function placementPoliciesUpdate() {
     ],
     state: "Disabled",
     vmMembers: [
-      "/subscriptions/{subscription-id}/resourceGroups/group1/providers/Microsoft.AVS/privateClouds/cloud1/clusters/cluster1/virtualMachines/vm-128",
-      "/subscriptions/{subscription-id}/resourceGroups/group1/providers/Microsoft.AVS/privateClouds/cloud1/clusters/cluster1/virtualMachines/vm-256",
+      "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/group1/providers/Microsoft.AVS/privateClouds/cloud1/clusters/cluster1/virtualMachines/vm-128",
+      "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/group1/providers/Microsoft.AVS/privateClouds/cloud1/clusters/cluster1/virtualMachines/vm-256",
     ],
   };
   const credential = new DefaultAzureCredential();