+++ sdk/cognitivelanguage/ai-language-text/test/public/analyzeBatch.spec.ts
@@ -13,7 +13,7 @@ import { AuthMethod, createClient, startRecorder } from "./utils/recordedClient"
 import { Context, Suite } from "mocha";
 import { Recorder, assertEnvironmentVariable, isPlaybackMode } from "@azure-tools/test-recorder";
 import { assert, matrix } from "@azure/test-utils";
-import { assertActionResults, assertRestError } from "./utils/resultHelper";
+import { assertActionsResults, assertRestError } from "./utils/resultHelper";
 import {
   expectation1,
   expectation10,
@@ -43,6 +43,11 @@ import {
 } from "./expectations";
 import { getDocIDsFromState } from "../../src/lro";
 
+const FIXME1 = {
+  // Remove this check when the service updates its message
+  excludedAdditionalProps: ["message"],
+};
+
 matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
   describe(`[${authMethod}] TextAnalysisClient`, function (this: Suite) {
     let recorder: Recorder;
@@ -50,8 +55,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
 
     beforeEach(async function (this: Context) {
       recorder = await startRecorder(this.currentTest);
-      client = createClient({
-        authMethod,
+      client = createClient(authMethod, {
         recorder,
       });
     });
@@ -91,7 +95,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
               }
             );
 
-            await assertActionResults(await poller.pollUntilDone(), expectation3);
+            await assertActionsResults(await poller.pollUntilDone(), expectation3);
           });
 
           it("key phrase extraction", async function () {
@@ -119,7 +123,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
                 updateIntervalInMs: pollingInterval,
               }
             );
-            await assertActionResults(await poller.pollUntilDone(), expectation5);
+            await assertActionsResults(await poller.pollUntilDone(), expectation5);
           });
 
           it("entity linking", async function () {
@@ -140,7 +144,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
                 updateIntervalInMs: pollingInterval,
               }
             );
-            await assertActionResults(await poller.pollUntilDone(), expectation6);
+            await assertActionsResults(await poller.pollUntilDone(), expectation6);
           });
 
           it("pii entity recognition", async function () {
@@ -162,7 +166,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
                 updateIntervalInMs: pollingInterval,
               }
             );
-            await assertActionResults(await poller.pollUntilDone(), expectation7);
+            await assertActionsResults(await poller.pollUntilDone(), expectation7);
           });
 
           it("pii entity recognition with filtered categories", async function () {
@@ -184,7 +188,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
                 updateIntervalInMs: pollingInterval,
               }
             );
-            await assertActionResults(await poller.pollUntilDone(), expectation8);
+            await assertActionsResults(await poller.pollUntilDone(), expectation8);
           });
 
           it("pii entity recognition with phi domain", async function () {
@@ -206,7 +210,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
                 updateIntervalInMs: pollingInterval,
               }
             );
-            await assertActionResults(await poller.pollUntilDone(), expectation24);
+            await assertActionsResults(await poller.pollUntilDone(), expectation24);
           });
 
           it("sentiment analysis with opinion mining", async function () {
@@ -233,7 +237,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
                 updateIntervalInMs: pollingInterval,
               }
             );
-            await assertActionResults(await poller.pollUntilDone(), expectation9);
+            await assertActionsResults(await poller.pollUntilDone(), expectation9);
           });
 
           it("healthcare", async function () {
@@ -254,84 +258,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
                 updateIntervalInMs: pollingInterval,
               }
             );
-            await assertActionResults(await poller.pollUntilDone(), expectation20);
-          });
-        });
-
-        describe("custom", function () {
-          it("entity recognition", async function () {
-            const docs = [
-              "A recent report by the Government Accountability Office (GAO) found that the dramatic increase in oil and natural gas development on federal lands over the past six years has stretched the staff of the BLM to a point that it has been unable to meet its environmental protection responsibilities.",
-            ];
-            const poller = await client.beginAnalyzeBatch(
-              [
-                {
-                  kind: AnalyzeBatchActionNames.CustomEntityRecognition,
-                  deploymentName: assertEnvironmentVariable(
-                    "LANGUAGE_CUSTOM_ENTITY_RECOGNITION_DEPLOYMENT_NAME"
-                  ),
-                  projectName: assertEnvironmentVariable(
-                    "LANGUAGE_CUSTOM_ENTITY_RECOGNITION_PROJECT_NAME"
-                  ),
-                },
-              ],
-              docs,
-              "en",
-              {
-                updateIntervalInMs: pollingInterval,
-              }
-            );
-            await assertActionResults(await poller.pollUntilDone(), expectation1);
-          });
-
-          it("single label classification action", async function () {
-            const docs = [
-              "A recent report by the Government Accountability Office (GAO) found that the dramatic increase in oil and natural gas development on federal lands over the past six years has stretched the staff of the BLM to a point that it has been unable to meet its environmental protection responsibilities.",
-            ];
-            const poller = await client.beginAnalyzeBatch(
-              [
-                {
-                  kind: AnalyzeBatchActionNames.CustomSingleLabelClassification,
-                  deploymentName: assertEnvironmentVariable(
-                    "LANGUAGE_CUSTOM_SINGLE_LABEL_CLASSIFICATION_DEPLOYMENT_NAME"
-                  ),
-                  projectName: assertEnvironmentVariable(
-                    "LANGUAGE_CUSTOM_SINGLE_LABEL_CLASSIFICATION_PROJECT_NAME"
-                  ),
-                },
-              ],
-              docs,
-              "en",
-              {
-                updateIntervalInMs: pollingInterval,
-              }
-            );
-            await assertActionResults(await poller.pollUntilDone(), expectation2);
-          });
-
-          it("multi label classification action", async function () {
-            const docs = [
-              "A recent report by the Government Accountability Office (GAO) found that the dramatic increase in oil and natural gas development on federal lands over the past six years has stretched the staff of the BLM to a point that it has been unable to meet its environmental protection responsibilities.",
-            ];
-            const poller = await client.beginAnalyzeBatch(
-              [
-                {
-                  kind: AnalyzeBatchActionNames.CustomMultiLabelClassification,
-                  deploymentName: assertEnvironmentVariable(
-                    "LANGUAGE_CUSTOM_MULTI_LABEL_CLASSIFICATION_DEPLOYMENT_NAME"
-                  ),
-                  projectName: assertEnvironmentVariable(
-                    "LANGUAGE_CUSTOM_MULTI_LABEL_CLASSIFICATION_PROJECT_NAME"
-                  ),
-                },
-              ],
-              docs,
-              "en",
-              {
-                updateIntervalInMs: pollingInterval,
-              }
-            );
-            await assertActionResults(await poller.pollUntilDone(), expectation4);
+            await assertActionsResults(await poller.pollUntilDone(), expectation20);
           });
         });
       });
@@ -483,7 +410,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
                 updateIntervalInMs: pollingInterval,
               }
             );
-            await assertActionResults(await poller.pollUntilDone(), expectation21);
+            await assertActionsResults(await poller.pollUntilDone(), expectation21, FIXME1);
           });
         });
 
@@ -506,7 +433,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
               updateIntervalInMs: pollingInterval,
             }
           );
-          await assertActionResults(await poller.pollUntilDone(), expectation19);
+          await assertActionsResults(await poller.pollUntilDone(), expectation19);
         });
 
         it("some documents with errors and multiple actions", async function () {
@@ -540,7 +467,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
               updateIntervalInMs: pollingInterval,
             }
           );
-          await assertActionResults(await poller.pollUntilDone(), expectation10);
+          await assertActionsResults(await poller.pollUntilDone(), expectation10, FIXME1);
         });
 
         it("all documents with errors and multiple actions", async function () {
@@ -574,7 +501,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
               updateIntervalInMs: pollingInterval,
             }
           );
-          await assertActionResults(await poller.pollUntilDone(), expectation11);
+          await assertActionsResults(await poller.pollUntilDone(), expectation11);
         });
 
         it("output order is same as the input's one with multiple actions", async function () {
@@ -602,7 +529,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
               updateIntervalInMs: pollingInterval,
             }
           );
-          await assertActionResults(await poller.pollUntilDone(), expectation12);
+          await assertActionsResults(await poller.pollUntilDone(), expectation12);
         });
 
         it("out of order input IDs with multiple actions", async function () {
@@ -630,7 +557,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
               updateIntervalInMs: pollingInterval,
             }
           );
-          await assertActionResults(await poller.pollUntilDone(), expectation13);
+          await assertActionsResults(await poller.pollUntilDone(), expectation13);
         });
 
         it("statistics", async function () {
@@ -700,7 +627,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
               updateIntervalInMs: pollingInterval,
             }
           );
-          await assertActionResults(await poller.pollUntilDone(), expectation14);
+          await assertActionsResults(await poller.pollUntilDone(), expectation14);
         });
 
         it("whole batch with no language hint", async function () {
@@ -727,7 +654,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
               updateIntervalInMs: pollingInterval,
             }
           );
-          await assertActionResults(await poller.pollUntilDone(), expectation14);
+          await assertActionsResults(await poller.pollUntilDone(), expectation14);
         });
 
         it("whole batch input with a language hint", async function () {
@@ -753,7 +680,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
               updateIntervalInMs: pollingInterval,
             }
           );
-          await assertActionResults(await poller.pollUntilDone(), expectation15);
+          await assertActionsResults(await poller.pollUntilDone(), expectation15);
         });
 
         it("invalid language hint", async function () {
@@ -776,7 +703,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
               updateIntervalInMs: pollingInterval,
             }
           );
-          await assertActionResults(await poller.pollUntilDone(), expectation16);
+          await assertActionsResults(await poller.pollUntilDone(), expectation16);
         });
 
         it("paged results with custom page size", async function () {
@@ -798,7 +725,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
               updateIntervalInMs: pollingInterval,
             }
           );
-          await assertActionResults(await poller.pollUntilDone(), expectation17, {
+          await assertActionsResults(await poller.pollUntilDone(), expectation17, {
             maxPageSize: 10,
           });
         });
@@ -914,8 +841,8 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
           const rehydratedPoller = await client.restoreAnalyzeBatchPoller(serializedState, {
             updateIntervalInMs: pollingInterval,
           });
-          await assertActionResults(await rehydratedPoller.pollUntilDone(), expectation26);
-          await assertActionResults(await originalPoller.pollUntilDone(), expectation26);
+          await assertActionsResults(await rehydratedPoller.pollUntilDone(), expectation26);
+          await assertActionsResults(await originalPoller.pollUntilDone(), expectation26);
         });
 
         describe("stringIndexType", function () {
@@ -934,7 +861,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
                 updateIntervalInMs: pollingInterval,
               }
             );
-            await assertActionResults(await poller.pollUntilDone(), expectation18);
+            await assertActionsResults(await poller.pollUntilDone(), expectation18);
           });
 
           it("family emoji wit skin tone modifier with Utf16CodeUnit", async function () {
@@ -952,7 +879,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
                 updateIntervalInMs: pollingInterval,
               }
             );
-            await assertActionResults(await poller.pollUntilDone(), expectation22);
+            await assertActionsResults(await poller.pollUntilDone(), expectation22);
           });
 
           it("family emoji wit skin tone modifier with UnicodeCodePoint", async function () {
@@ -970,7 +897,99 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
                 updateIntervalInMs: pollingInterval,
               }
             );
-            await assertActionResults(await poller.pollUntilDone(), expectation23);
+            await assertActionsResults(await poller.pollUntilDone(), expectation23);
+          });
+        });
+      });
+    });
+  });
+});
+
+matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
+  describe(`[${authMethod}] TextAnalysisClient`, function (this: Suite) {
+    let recorder: Recorder;
+    let client: TextAnalysisClient;
+
+    beforeEach(async function (this: Context) {
+      recorder = await startRecorder(this.currentTest);
+      client = createClient(authMethod, {
+        resource: "CustomText",
+        recorder,
+      });
+    });
+
+    afterEach(async function () {
+      await recorder.stop();
+    });
+
+    describe("analyzeBatch", function () {
+      const pollingInterval = isPlaybackMode() ? 0 : 2000;
+
+      describe("actions", function () {
+        describe("custom", function () {
+          it("entity recognition", async function () {
+            const docs = [
+              "A recent report by the Government Accountability Office (GAO) found that the dramatic increase in oil and natural gas development on federal lands over the past six years has stretched the staff of the BLM to a point that it has been unable to meet its environmental protection responsibilities.",
+            ];
+            const poller = await client.beginAnalyzeBatch(
+              [
+                {
+                  kind: AnalyzeBatchActionNames.CustomEntityRecognition,
+                  deploymentName: assertEnvironmentVariable("CUSTOM_ENTITIES_DEPLOYMENT_NAME"),
+                  projectName: assertEnvironmentVariable("CUSTOM_ENTITIES_PROJECT_NAME"),
+                },
+              ],
+              docs,
+              "en",
+              {
+                updateIntervalInMs: pollingInterval,
+              }
+            );
+            await assertActionsResults(await poller.pollUntilDone(), expectation1);
+          });
+
+          it("single label classification action", async function () {
+            const docs = [
+              "A recent report by the Government Accountability Office (GAO) found that the dramatic increase in oil and natural gas development on federal lands over the past six years has stretched the staff of the BLM to a point that it has been unable to meet its environmental protection responsibilities.",
+            ];
+            const poller = await client.beginAnalyzeBatch(
+              [
+                {
+                  kind: AnalyzeBatchActionNames.CustomSingleLabelClassification,
+                  deploymentName: assertEnvironmentVariable(
+                    "SINGLE_LABEL_CLASSIFY_DEPLOYMENT_NAME"
+                  ),
+                  projectName: assertEnvironmentVariable("SINGLE_LABEL_CLASSIFY_PROJECT_NAME"),
+                },
+              ],
+              docs,
+              "en",
+              {
+                updateIntervalInMs: pollingInterval,
+              }
+            );
+            await assertActionsResults(await poller.pollUntilDone(), expectation2);
+          });
+
+          it("multi label classification action", async function () {
+            const docs = [
+              "A recent report by the Government Accountability Office (GAO) found that the dramatic increase in oil and natural gas development on federal lands over the past six years has stretched the staff of the BLM to a point that it has been unable to meet its environmental protection responsibilities.",
+            ];
+            const poller = await client.beginAnalyzeBatch(
+              [
+                {
+                  kind: AnalyzeBatchActionNames.CustomMultiLabelClassification,
+                  deploymentName: assertEnvironmentVariable("MULTI_LABEL_CLASSIFY_DEPLOYMENT_NAME"),
+                  projectName: assertEnvironmentVariable("MULTI_LABEL_CLASSIFY_PROJECT_NAME"),
+                },
+              ],
+              docs,
+              "en",
+              {
+                updateIntervalInMs: pollingInterval,
+              }
+            );
+            await assertActionsResults(await poller.pollUntilDone(), expectation4);
           });
         });
       });