+++ sdk/core/core-rest-pipeline/src/policies/tracingPolicy.ts
@@ -7,13 +7,13 @@ import {
   TracingSpan,
   createTracingClient,
 } from "@azure/core-tracing";
-import { SDK_VERSION } from "../constants";
-import { PipelineRequest, PipelineResponse, SendRequest } from "../interfaces";
-import { PipelinePolicy } from "../pipeline";
-import { getUserAgentValue } from "../util/userAgent";
-import { logger } from "../log";
+import { SDK_VERSION } from "../constants.js";
+import { PipelineRequest, PipelineResponse, SendRequest } from "../interfaces.js";
+import { PipelinePolicy } from "../pipeline.js";
+import { getUserAgentValue } from "../util/userAgent.js";
+import { logger } from "../log.js";
 import { getErrorMessage, isError } from "@azure/core-util";
-import { isRestError } from "../restError";
+import { isRestError } from "../restError.js";
 
 /**
  * The programmatic identifier of the tracingPolicy.