+++ sdk/monitor/monitor-opentelemetry-exporter/src/export/metric.ts
@@ -1,17 +1,17 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
-import { diag } from "@opentelemetry/api";
+import { context, diag } from "@opentelemetry/api";
 import {
   AggregationTemporality,
   InstrumentType,
   PushMetricExporter,
   ResourceMetrics,
 } from "@opentelemetry/sdk-metrics";
-import { ExportResult, ExportResultCode } from "@opentelemetry/core";
+import { ExportResult, ExportResultCode, suppressTracing } from "@opentelemetry/core";
 import { AzureMonitorBaseExporter } from "./base";
-import { AzureMonitorExporterOptions } from "../config";
 import { TelemetryItem as Envelope } from "../generated";
 import { resourceMetricsToEnvelope } from "../utils/metricUtils";
+import { AzureMonitorExporterOptions } from "../config";
 
 /**
  * Azure Monitor OpenTelemetry Metric Exporter.
@@ -33,6 +33,7 @@ export class AzureMonitorMetricExporter
    * Initializes a new instance of the AzureMonitorMetricExporter class.
    * @param AzureExporterConfig - Exporter configuration.
    */
+
   constructor(options: AzureMonitorExporterOptions = {}) {
     super(options);
     this._aggregationTemporality = AggregationTemporality.CUMULATIVE;
@@ -56,7 +57,10 @@ export class AzureMonitorMetricExporter
     diag.info(`Exporting ${metrics.scopeMetrics.length} metrics(s). Converting to envelopes...`);
 
     let envelopes: Envelope[] = resourceMetricsToEnvelope(metrics, this._instrumentationKey);
-    resultCallback(await this._exportEnvelopes(envelopes));
+    // Supress tracing until OpenTelemetry Metrics SDK support it
+    context.with(suppressTracing(context.active()), async () => {
+      resultCallback(await this._exportEnvelopes(envelopes));
+    });
   }
 
   /**