+++ sdk/monitor/monitor-opentelemetry-exporter/src/index.ts
@@ -1,9 +1,11 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+export { ApplicationInsightsSampler } from "./sampling";
 export { AzureMonitorBaseExporter } from "./export/base";
 export { AzureMonitorTraceExporter } from "./export/trace";
 export { AzureMonitorMetricExporter } from "./export/metric";
+export { _AzureMonitorStatsbeatExporter } from "./export/statsbeat";
 export { AzureMonitorExporterOptions } from "./config";
 export { ServiceApiVersion } from "./Declarations/Constants";
 export {