+++ sdk/notificationhubs/notification-hubs/src/client/deleteInstallation.ts
@@ -22,7 +22,7 @@ export function deleteInstallation(
   options: OperationOptions = {}
 ): Promise<NotificationHubsResponse> {
   return tracingClient.withSpan(
-    `NotificationHubsClientContext-${OPERATION_NAME}`,
+    `NotificationHubsClientContext.${OPERATION_NAME}`,
     options,
     async (updatedOptions) => {
       const endpoint = context.requestUrl();