+++ sdk/cosmosdb/cosmos/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 3.17.2 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 3.17.1 (2022-09-12)
 
 ### Bugs Fixed