+++ sdk/monitor/monitor-opentelemetry-exporter/CHANGELOG.md
@@ -1,12 +1,14 @@
 # Release History
 
-## 1.0.0-beta.9 (Unreleased)
+## 1.0.0-beta.9 (2022-10-20)
 
 ### Features Added
 
+- Application Insights Sampler.
 - Added retriable behavior for 502, 503 and 504 status codes.
 - Export Metric attributes and Histogram Min/Max values.
 - Added new config options disableOfflineStorage, storageDirectory and exposed ApplicationInsightsClientOptionalParams for HTTP client extra configuration.
+- Added Network Statsbeat Metrics.
 
 ### Breaking Changes
 
@@ -15,7 +17,8 @@
 
 ### Bugs Fixed
 
-### Other Changes
+- Suppress tracing while exporting metrics.
+- Envelopes not populating sampleRate correctly.
 
 ## 1.0.0-beta.8 (2022-07-07)